requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.7
Pillow==10.0.1